Cpu optimization (#44)
* Remove Boost warnings


* Optimize Pandar40P


* Pandar64 optimization


* PointTypes. standardize point types


* Add radian based elevation calculation Pandar40P,64


* qt64. radian based coordinate calculation


* xt32. radian based coordinate calculation


* qt128. radian based coordinate calculation


* qt128. radian based coordinate calculation


* pandar40. precalculate sin/cos


* pandarxt. precalculate te sin/cos


* apply sin/cos optimization to all decoders


* launch. fix python launch file


* point_type. caetr SSE alignment optimization


* hesai_hw. remove warning


* point_types. add sse description for caetr


* README. run tests only for nebula packages


* clang. apply ament_clang reformat


* Replace 40P, 64, QT64 and XT32s' PCD files to fix testing

* remove clang-format from tests


* Local path check not required for online acquisition

* Adding timeouts to correction data acquisition

* Disable calibration query


---------

+++ nebula_decoders/include/nebula_decoders/nebula_decoders_hesai/decoders/hesai_scan_decoder.hpp
@@ -26,7 +26,7 @@ protected:
   bool has_scanned_{};
   double dual_return_distance_threshold_{};
   uint32_t first_timestamp_{};
-  double first_timestamp_tmp{};
+  uint32_t first_timestamp_tmp{};
 
   /// @brief SensorConfiguration for this decoder
   std::shared_ptr<drivers::HesaiSensorConfiguration> sensor_configuration_;
